backend/requirements.txt
flask
python-docx
PyPDF2
scikit-learn
nltk
